feat(alist_v3): add PassIPToUpsteam configuration option
If the PassIPToUpsteam option is enabled, the client's IP address will be
forwarded to the upstream alist server.

This change adds the necessary header
X-Forwarded-For to the HTTP request when the condition is met.

This can help the upstream alist server to correctly get the download link by the downloader operator's ip address.